+++ stitchee/__init__.py
@@ -1,0 +1,5 @@
+"""Package version."""
+
+from importlib.metadata import version
+
+__version__ = version("stitchee")