+++ stitchee/harmony/service_adapter.py
@@ -12,15 +12,15 @@ from harmony_service_lib.util import bbox_to_geometry, stage
 from pystac import Item
 from pystac.item import Asset
 
-from concatenator.attribute_handling import construct_history, retrieve_history
-from concatenator.harmony.download_worker import multi_core_download
-from concatenator.harmony.util import (
+from stitchee.concatenate import concatenate
+from stitchee.harmony.download_worker import multi_core_download
+from stitchee.harmony.util import (
     _get_netcdf_urls,
     _get_output_bounding_box,
     _get_output_date_range,
     sizeof_fmt,
 )
-from concatenator.stitchee import stitchee
+from stitchee.history_handling import construct_history, retrieve_history
 
 
 class StitcheeAdapter(BaseHarmonyAdapter):
@@ -117,11 +117,9 @@ class StitcheeAdapter(BaseHarmonyAdapter):
                 output_path = str(Path(temp_dir).joinpath(filename).resolve())
 
                 # # --- Run STITCHEE ---
-                stitchee(
+                concatenate(
                     [str(f) for f in input_files],
                     output_path,
-                    write_tmp_flat_concatenated=False,
-                    keep_tmp_files=False,
                     concat_dim="mirror_step",  # This is currently set only for TEMPO
                     sorting_variable="geolocation/time",  # This is currently set only for TEMPO
                     history_to_append=new_history_json,