+++ tests/integration/test_concat_with_subsetting_first.py
@@ -2,7 +2,7 @@ import datetime as dt
 
 from harmony import BBox, Client, Collection, Request
 
-from concatenator.stitchee import stitchee
+from stitchee.concatenate import concatenate
 
 
 def test_concat_with_subsetting_first(temp_output_dir):
@@ -17,6 +17,7 @@ def test_concat_with_subsetting_first(temp_output_dir):
             "stop": dt.datetime(2024, 5, 13, 20, 0, 0),
         },
         spatial=BBox(-130, 30, -115, 35),
+        extend=False,
         concatenate=False,
     )
     if not request.is_valid():
@@ -33,15 +34,13 @@ def test_concat_with_subsetting_first(temp_output_dir):
     print(f"File names: {file_names}")
 
     # Try concatenating the resulting files
-    output_path = stitchee(
+    output_path = concatenate(
         file_names,
         output_file=str(
             (temp_output_dir / "output_harmony_subsetting_to_stitchee_test.nc").resolve()
         ),
         concat_dim="mirror_step",
         concat_method="xarray-concat",
-        write_tmp_flat_concatenated=True,
-        keep_tmp_files=True,
     )
 
     assert output_path